[proliferate] split diff viewer components
