+++ desktop/src/components/ui/content/diff/ChatDiffViewer.tsx
@@ -1,0 +1,263 @@
+import { useMemo, useState, type CSSProperties } from "react";
+import { DiffLineContent } from "@/components/ui/content/diff/DiffLineContent";
+import { useResolvedMode } from "@/hooks/theme/derived/use-resolved-mode";
+import type {
+  CollapsedContext,
+  DiffLine,
+  ParsedPatch,
+} from "@/lib/domain/files/diff-parser";
+import type { HighlightedToken } from "@/lib/infra/editor/highlighting";
+
+type ChatDiffRow =
+  | { kind: "line"; key: string; line: DiffLine }
+  | { kind: "collapsed"; key: string; section: CollapsedContext };
+
+const CHAT_DIFF_PRE_STYLE = {
+  color: "var(--diffs-fg)",
+  backgroundColor: "var(--diffs-bg)",
+  "--diffs-bg": "var(--codex-diffs-surface)",
+  "--diffs-addition-color": "var(--diffs-addition-color-override)",
+  "--diffs-deletion-color": "var(--diffs-deletion-color-override)",
+  "--diffs-min-number-column-width": "4ch",
+  "--diffs-min-number-column-width-default": "3ch",
+} as CSSProperties;
+
+const CHAT_DIFF_CODE_BASE_STYLE = {
+  "--diffs-column-content-width": "700px",
+  "--diffs-column-width": "736px",
+} as CSSProperties;
+
+function getChatLineType(line: DiffLine): "context" | "change-addition" | "change-deletion" {
+  switch (line.type) {
+    case "added":
+      return "change-addition";
+    case "removed":
+      return "change-deletion";
+    case "context":
+      return "context";
+  }
+}
+
+function getChatLineNumber(line: DiffLine): number | null {
+  if (line.type === "removed") {
+    return line.oldLineNum;
+  }
+  return line.newLineNum ?? line.oldLineNum;
+}
+
+function getChatLineIndex(line: DiffLine): string {
+  return `${line.oldLineNum ?? ""},${line.newLineNum ?? ""}`;
+}
+
+function getChatRows(
+  parsed: ParsedPatch,
+  expandedCollapsedKeys: Set<string>,
+): ChatDiffRow[] {
+  const rows: ChatDiffRow[] = [];
+
+  parsed.hunks.forEach((hunk, hunkIndex) => {
+    hunk.items.forEach((item, itemIndex) => {
+      const key = `${hunkIndex}-${itemIndex}`;
+      if ("kind" in item && item.kind === "collapsed") {
+        if (expandedCollapsedKeys.has(key)) {
+          item.lines.forEach((line) => {
+            rows.push({
+              kind: "line",
+              key: `${key}-${line.tokenIndex}`,
+              line,
+            });
+          });
+        } else {
+          rows.push({ kind: "collapsed", key, section: item });
+        }
+        return;
+      }
+
+      const line = item as DiffLine;
+      rows.push({
+        kind: "line",
+        key: `${key}-${line.tokenIndex}`,
+        line,
+      });
+    });
+  });
+
+  return rows;
+}
+
+function getChatLineNumberDigits(rows: ChatDiffRow[]): number {
+  let maxLineNumber = 0;
+  for (const row of rows) {
+    if (row.kind !== "line") continue;
+    maxLineNumber = Math.max(maxLineNumber, getChatLineNumber(row.line) ?? 0);
+  }
+  return Math.max(String(maxLineNumber).length, 1);
+}
+
+function ChatGutterCell({ line }: { line: DiffLine }) {
+  const lineType = getChatLineType(line);
+  const lineNumber = getChatLineNumber(line);
+
+  return (
+    <div
+      data-line-type={lineType}
+      data-column-number={lineNumber ?? undefined}
+      data-line-index={getChatLineIndex(line)}
+      className="diff-gutter-cell box-border flex min-h-[var(--diffs-line-height)] w-[var(--diffs-column-number-width)] min-w-[var(--diffs-column-number-width)] items-center justify-end px-2 text-right tabular-nums"
+    >
+      <span data-line-number-content="">{lineNumber ?? ""}</span>
+    </div>
+  );
+}
+
+function ChatContentCell({
+  line,
+  tokens,
+}: {
+  line: DiffLine;
+  tokens: HighlightedToken[][] | null;
+}) {
+  const lineType = getChatLineType(line);
+  const lineNumber = getChatLineNumber(line);
+  const altLineNumber = line.type === "context" ? line.oldLineNum : undefined;
+
+  return (
+    <div
+      data-line={lineNumber ?? undefined}
+      data-alt-line={altLineNumber}
+      data-line-type={lineType}
+      data-line-index={getChatLineIndex(line)}
+      className="diff-content-cell relative flex min-h-[var(--diffs-line-height)] min-w-max items-center pr-3 pl-2 whitespace-pre"
+    >
+      <DiffLineContent line={line} tokens={tokens} />
+    </div>
+  );
+}
+
+function ChatCollapsedRow({
+  section,
+  onExpand,
+}: {
+  section: CollapsedContext;
+  onExpand: () => void;
+}) {
+  return (
+    <button
+      type="button"
+      data-separator="simple"
+      onClick={onExpand}
+      aria-label={`Show ${section.lineCount} unmodified lines`}
+      title={`${section.lineCount} unmodified lines`}
+      className="diff-content-cell flex min-h-[var(--diffs-line-height)] cursor-pointer items-center border-0 bg-transparent p-0 text-left font-[inherit] text-[inherit] leading-[inherit]"
+    />
+  );
+}
+
+export function ChatDiffViewer({
+  parsed,
+  tokens,
+  className,
+  viewportClassName,
+}: {
+  parsed: ParsedPatch;
+  tokens: HighlightedToken[][] | null;
+  className?: string;
+  viewportClassName?: string;
+}) {
+  const resolvedMode = useResolvedMode();
+  const [expandedCollapsedKeys, setExpandedCollapsedKeys] = useState<Set<string>>(
+    new Set(),
+  );
+  const rows = useMemo(
+    () => getChatRows(parsed, expandedCollapsedKeys),
+    [parsed, expandedCollapsedKeys],
+  );
+  const rowSpan = Math.max(rows.length, 1);
+  const lineNumberDigits = getChatLineNumberDigits(rows);
+  const codeStyle = useMemo(
+    () => ({
+      ...CHAT_DIFF_CODE_BASE_STYLE,
+      "--diffs-column-number-width": `max(24px, ${lineNumberDigits + 1}ch)`,
+    }) as CSSProperties,
+    [lineNumberDigits],
+  );
+
+  const expandCollapsedRow = (key: string) => {
+    setExpandedCollapsedKeys((prev) => {
+      const next = new Set(prev);
+      next.add(key);
+      return next;
+    });
+  };
+
+  return (
+    <div className={className ?? ""}>
+      <div
+        style={{ overscrollBehavior: "none" }}
+        className={`relative [contain:content] composer-diff-simple-line overflow-x-auto overflow-y-auto ${
+          viewportClassName ?? ""
+        }`}
+      >
+        <pre
+          data-diff=""
+          data-theme-type={resolvedMode === "dark" ? "dark" : "light"}
+          data-indicators="bars"
+          data-background=""
+          data-diff-type="single"
+          data-overflow="scroll"
+          data-interactive-lines=""
+          data-interactive-line-numbers=""
+          tabIndex={0}
+          style={CHAT_DIFF_PRE_STYLE}
+          className="m-0 w-full min-w-max bg-[var(--codex-diffs-surface)] p-0 font-[family:var(--diffs-font-family)] text-[length:var(--diffs-font-size)] leading-[var(--diffs-line-height)] text-[color:var(--diffs-fg)]"
+        >
+          <code
+            data-code=""
+            data-unified=""
+            style={codeStyle}
+            className="grid grid-cols-[var(--diffs-column-number-width)_minmax(max-content,1fr)]"
+          >
+            <div
+              data-gutter=""
+              style={{ gridRow: `span ${rowSpan}` }}
+              className="sticky left-0 z-10 grid auto-rows-[var(--diffs-line-height)] bg-[var(--diffs-bg)]"
+            >
+              {rows.map((row) =>
+                row.kind === "line" ? (
+                  <ChatGutterCell key={`g-${row.key}`} line={row.line} />
+                ) : (
+                  <div
+                    key={`g-${row.key}`}
+                    data-separator="simple"
+                    className="diff-gutter-cell min-h-[var(--diffs-line-height)]"
+                  />
+                ),
+              )}
+            </div>
+            <div
+              data-content=""
+              style={{ gridRow: `span ${rowSpan}` }}
+              className="grid auto-rows-[var(--diffs-line-height)]"
+            >
+              {rows.map((row) =>
+                row.kind === "line" ? (
+                  <ChatContentCell
+                    key={`c-${row.key}`}
+                    line={row.line}
+                    tokens={tokens}
+                  />
+                ) : (
+                  <ChatCollapsedRow
+                    key={`c-${row.key}`}
+                    section={row.section}
+                    onExpand={() => expandCollapsedRow(row.key)}
+                  />
+                ),
+              )}
+            </div>
+          </code>
+        </pre>
+      </div>
+    </div>
+  );
+}
